# docmd/archives-modules/_archived/immobilisation-legacy-preP0/openapi/immobilisation.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SPOFE — Immobilisation API",
    "description": "## Contractual API — Immobilisation Module v1.0.0\n\n### Overview\nThis API exposes the Immobilisation (Fixed Assets) module for asset management, depreciation tracking, maintenance, and renewals.\nIt follows the **CQRS Read-Only** architecture pattern.\n\n### Contracts\n- **IMM-CS-DEP-01**: Cost-Structure - Depreciation summary per period\n- **IMM-CS-DEP-02**: Cost-Structure - Depreciation export with allocations\n- **IMM-CS-ALL-01**: Cost-Structure - Effective allocations\n- **IMM-CS-MNT-01**: Cost-Structure - Maintenance summary per asset\n- **IMM-CS-MNT-02**: Cost-Structure - Maintenance by period\n- **IMM-BUD-REN-01**: Budget - Renewal projections\n- **IMM-BUD-REN-02**: Budget - Renewals by year\n- **IMM-BUD-MNT-01**: Budget - Maintenance costs (OPEX)\n- **IMM-BUD-DEP-01**: Budget - Depreciation summary\n\n### Multi-Tenant\nAll endpoints require the `X-Tenant-Id` header for tenant isolation.\n\n### Versioning\n- API Version: 1.0.0\n- Contract Version: IMMOBILISATION_CONTRACT v1.0.0\n\n### Architecture\n- ✅ GET endpoints only (read-only)\n- ✅ 1 endpoint = 1 read-model SQL\n- ✅ Zero business logic in controllers\n- ✅ Zero Guardian validation on read\n\nGenerated: 2026-02-01T17:01:01.949Z",
    "version": "1.0.0",
    "contact": {
      "name": "SPOFE Team",
      "url": "https://spofe.example.com",
      "email": "api@spofe.example.com"
    },
    "license": {
...
    }
  },
  "servers": [
...
  "paths": {
...
    "/immobilisation/depreciation/summary": {
      "get": {
        "operationId": "Immobilisation_getDepreciationSummary",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get depreciation summary [IMM-CS-DEP-01]",
        "description": "CONTRACTUAL: Cost-Structure. Returns aggregated depreciation for a period. Read-model: rm_asset_depreciation_summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "description": "Period (YYYY-MM)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "example": "2025-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Depreciation summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepreciationSummaryDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/immobilisation/allocations": {
      "get": {
        "operationId": "Immobilisation_listAllocations",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "List effective allocations [IMM-CS-ALL-01]",
        "description": "CONTRACTUAL: Cost-Structure. Returns effective allocations. Read-model: rm_asset_allocation_effective",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "name": "assetId",
            "in": "query",
            "description": "Filter by asset ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "description": "Filter by target type",
            "schema": {
              "type": "string",
              "enum": [
                "PRODUCT",
                "SERVICE",
                "PROJECT"
              ]
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "description": "Filter by target ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated allocations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAllocationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/immobilisation/maintenance/summary": {
      "get": {
        "operationId": "Immobilisation_getMaintenanceSummary",
        "tags": [
          "Immobilisation — Cost-Structure Contract",
          "Immobilisation — Budget Contract"
        ],
        "summary": "Get maintenance summary [IMM-CS-MNT-01, IMM-BUD-MNT-01]",
        "description": "CONTRACTUAL: Cost-Structure, Budget. Returns maintenance costs per asset. Read-model: rm_asset_maintenance_summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "assetId",
            "in": "query",
            "description": "Filter by asset ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceSummaryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/immobilisation/maintenance/by-period": {
      "get": {
        "operationId": "Immobilisation_getMaintenanceByPeriod",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get maintenance by period [IMM-CS-MNT-02]",
        "description": "CONTRACTUAL: Cost-Structure. Returns maintenance costs by period. Read-model: rm_asset_maintenance_by_period",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "fromPeriod",
            "in": "query",
            "description": "Start period (YYYY-MM)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          },
          {
            "name": "toPeriod",
            "in": "query",
            "description": "End period (YYYY-MM)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance by period",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceByPeriodResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/immobilisation/cost-structure/depreciation/summary": {
      "get": {
        "operationId": "ImmobilisationCostStructure_getDepreciationSummary",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get depreciation summary [IMM-CS-DEP-01]",
        "description": "CONTRACTUAL: Cost-Structure dedicated endpoint",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "description": "Period (YYYY-MM)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Depreciation summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepreciationSummaryDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/immobilisation/cost-structure/depreciation/export": {
      "get": {
        "operationId": "ImmobilisationCostStructure_getDepreciationExport",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get depreciation export [IMM-CS-DEP-02]",
        "description": "CONTRACTUAL: Cost-Structure dedicated endpoint",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "description": "Period (YYYY-MM)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Depreciation export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepreciationCostStructureExportDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/immobilisation/cost-structure/allocations": {
      "get": {
        "operationId": "ImmobilisationCostStructure_getAllocations",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get allocations [IMM-CS-ALL-01]",
        "description": "CONTRACTUAL: Cost-Structure dedicated endpoint",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "assetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PRODUCT",
                "SERVICE",
                "PROJECT"
              ]
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allocations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AllocationItemDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/immobilisation/cost-structure/maintenance/summary": {
      "get": {
        "operationId": "ImmobilisationCostStructure_getMaintenanceSummary",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get maintenance summary [IMM-CS-MNT-01]",
        "description": "CONTRACTUAL: Cost-Structure dedicated endpoint",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "assetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceSummaryItemDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/immobilisation/cost-structure/maintenance/by-period": {
      "get": {
        "operationId": "ImmobilisationCostStructure_getMaintenanceByPeriod",
        "tags": [
          "Immobilisation — Cost-Structure Contract"
        ],
        "summary": "Get maintenance by period [IMM-CS-MNT-02]",
        "description": "CONTRACTUAL: Cost-Structure dedicated endpoint",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdHeader"
          },
          {
            "name": "fromPeriod",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          },
          {
            "name": "toPeriod",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance by period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceByPeriodItemDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "TenantIdHeader": {
        "name": "X-Tenant-Id",
        "in": "header",
        "required": true,
        "description": "Tenant ID for multi-tenant isolation",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "PageQuery": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-indexed)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1,
          "example": 1
        }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "description": "Items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": 20
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string",
            "example": "/api/immobilisation/assets"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "DepreciationSummaryDTO": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "example": "2025-01"
          },
          "totalDepreciation": {
            "type": "number",
            "example": 4500000
          },
          "assetCount": {
            "type": "integer",
            "example": 12
          },
          "currency": {
            "type": "string",
            "example": "XAF"
          }
        },
        "required": [
          "period",
          "totalDepreciation",
          "assetCount",
          "currency"
        ]
      },
      "DepreciationCostStructureExportDTO": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "example": "2025-01"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepreciationCostStructureExportItemDTO"
            }
          },
          "total": {
            "type": "integer",
            "example": 5
          }
        },
        "required": [
          "period",
          "items",
          "total"
        ]
      },
      "DepreciationCostStructureExportItemDTO": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "example": "asset-001"
          },
          "depreciationAmount": {
            "type": "number",
            "example": 183333.33
          },
          "targetType": {
            "type": "string",
            "enum": [
              "PRODUCT",
              "SERVICE",
              "PROJECT"
            ],
            "example": "PRODUCT"
          },
          "targetId": {
            "type": "string",
            "example": "product-001"
          },
          "percentage": {
            "type": "number",
            "example": 60
          },
          "allocatedAmount": {
            "type": "number",
            "example": 110000
          },
          "currency": {
            "type": "string",
            "example": "XAF"
          }
        },
        "required": [
          "assetId",
          "depreciationAmount",
          "targetType",
          "targetId",
          "percentage",
          "allocatedAmount",
          "currency"
        ]
      },
      "PaginatedAllocationResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllocationItemDTO"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total",
          "hasMore"
        ]
      },
      "AllocationItemDTO": {
        "type": "object",
        "properties": {
          "allocationId": {
            "type": "string",
            "example": "alloc-001"
          },
          "assetId": {
            "type": "string",
            "example": "asset-001"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "PRODUCT",
              "SERVICE",
              "PROJECT"
            ],
            "example": "PRODUCT"
          },
          "targetId": {
            "type": "string",
            "example": "product-001"
          },
          "percentage": {
            "type": "number",
            "example": 60
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date",
            "example": "2024-01-01"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date",
            "example": null
          }
        },
        "required": [
          "allocationId",
          "assetId",
          "targetType",
          "targetId",
          "percentage",
          "effectiveFrom"
        ]
      },
...
      "MaintenanceSummaryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaintenanceSummaryItemDTO"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "MaintenanceSummaryItemDTO": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "example": "asset-001"
          },
          "totalMaintenanceCost": {
            "type": "number",
            "example": 500000
          },
          "interventionCount": {
            "type": "integer",
            "example": 3
          },
          "firstIntervention": {
            "type": "string",
            "format": "date",
            "example": "2024-06-15"
          },
          "lastIntervention": {
            "type": "string",
            "format": "date",
            "example": "2025-01-10"
          },
          "currency": {
            "type": "string",
            "example": "XAF"
          }
        },
        "required": [
          "assetId",
          "totalMaintenanceCost",
          "interventionCount",
          "currency"
        ]
      },
      "MaintenanceByPeriodResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaintenanceByPeriodItemDTO"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "MaintenanceByPeriodItemDTO": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "example": "2025-01"
          },
          "totalMaintenanceCost": {
            "type": "number",
            "example": 225000
          },
          "interventionCount": {
            "type": "integer",
            "example": 5
          },
          "assetsMaintained": {
            "type": "integer",
            "example": 3
          },
          "currency": {
            "type": "string",
            "example": "XAF"
          }
        },
        "required": [
          "period",
          "totalMaintenanceCost",
          "interventionCount",
          "assetsMaintained",
          "currency"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "tenantId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tenantId": []
    }
  ]
}